+++ source/modules/sis3316.cpp
@@ -10,12 +10,12 @@ using std::to_string;
 
 void SIS3316::reset_raw_leaves() {
     for (size_t n_leaf = 0; n_leaf < 8; ++n_leaf) {
-        amplitude_e1.leaves[n_leaf] = numeric_limits<unsigned int>::quiet_NaN();
-        amplitude_e2.leaves[n_leaf] = numeric_limits<unsigned int>::quiet_NaN();
-        time.leaves[n_leaf] = numeric_limits<unsigned int>::quiet_NaN();
-        timestamp.leaves[n_leaf] = numeric_limits<unsigned int>::quiet_NaN();
+        amplitude_e1.leaves[n_leaf] = numeric_limits<double>::quiet_NaN();
+        amplitude_e2.leaves[n_leaf] = numeric_limits<double>::quiet_NaN();
+        time.leaves[n_leaf] = numeric_limits<double>::quiet_NaN();
+        timestamp.leaves[n_leaf] = numeric_limits<double>::quiet_NaN();
     }
-    time_RF.leaves[0] = numeric_limits<unsigned int>::quiet_NaN();
+    time_RF.leaves[0] = numeric_limits<double>::quiet_NaN();
 }
 
 void SIS3316::set_up_raw_branches_for_reading(TTree *tree) {
